Add detail view controller that implements protocol
Note the use of key-value coding and the key-path
map for syncing fields with the data source. Note
we're provide extra code for the common niceties
of a form, mainly 'inline' editing.

Our view controllers and views should manage a
data source and be able to work with custom nib
names for them and their child views.